Admin access to telebot commands

+++ src/telebot_messages.py
@@ -44,43 +44,64 @@ def start_bot():
     def on_chat_message(msg):
         content_type, chat_type, chat_id = telepot.glance(msg)
 
+        # obtenemos el comando sudo
         pass_file = open(os.path.join(parent_folder, "etc", "sys_pass.txt"), "r")
         sys_pass = pass_file.readline()
         print(sys_pass)
         pass_file.close()
+
+        # cargamos los IDs de administradores
+        ids_file = open(os.path.join(parent_folder, "etc", "telegram_ids.cfg"), "r")
+        admin_ids = ids_file.read()
+        ids_file.close()
+        str_chat_id = str(chat_id)
+
+        no_rights = False
+
         sudo = "echo \"" + sys_pass + "\" | sudo -S " # es importante que haya un espacio despues de -S
         if content_type == "text":
             text = msg["text"]
+
             if text == "/restart_nuc":
-                my_bot.sendMessage(chat_id=chat_id, text="Restarting NUC...")
-                os.system(sudo + "shutdown -r 0")
+                if str_chat_id in admin_ids:
+                    my_bot.sendMessage(chat_id=chat_id, text="Restarting NUC...")
+                    os.system(sudo + "shutdown -r 0")
+                else:
+                    no_rights = True
 
             elif text == "/start_ngrok":
-                my_bot.sendMessage(chat_id=chat_id, text="Starting Ngrok service...")
-                ret = os.system(sudo + "service server_ngrok start")
-                if ret == 0:
-                    my_bot.sendMessage(chat_id=chat_id, text=emojize(':thumbs_up:'))
+                if str_chat_id in admin_ids:
+                    my_bot.sendMessage(chat_id=chat_id, text="Starting Ngrok service...")
+                    ret = os.system(sudo + "service server_ngrok start")
+                    if ret == 0:
+                        my_bot.sendMessage(chat_id=chat_id, text=emojize(':thumbs_up:'))
+                    else:
+                        my_bot.sendMessage(chat_id=chat_id, text='fail')
                 else:
-                    my_bot.sendMessage(chat_id=chat_id, text='fail')
+                 no_rights = True
 
             elif text == "/stop_ngrok":
-                my_bot.sendMessage(chat_id=chat_id, text="Stopping Ngrok service...")
-                ret = os.system(sudo + "service server_ngrok stop")
-                if ret == 0:
-                    my_bot.sendMessage(chat_id=chat_id, text=emojize(':thumbs_up:'))
+                if str_chat_id in admin_ids:
+                    my_bot.sendMessage(chat_id=chat_id, text="Stopping Ngrok service...")
+                    ret = os.system(sudo + "service server_ngrok stop")
+                    if ret == 0:
+                        my_bot.sendMessage(chat_id=chat_id, text=emojize(':thumbs_up:'))
+                    else:
+                        my_bot.sendMessage(chat_id=chat_id, text='fail')
                 else:
-                    my_bot.sendMessage(chat_id=chat_id, text='fail')
+                    no_rights = True
 
             elif text == "/status":
                 my_bot.sendMessage(chat_id=chat_id, text=str(get_saved_info()))
 
-            elif "/" in text:
-                my_bot.sendMessage(chat_id=chat_id, text="There's nothing to do here... ")
-                
             else:
-                my_bot.sendMessage(chat_id=chat_id, text="/restart_nuc\n/start_ngrok\n/stop_ngrok\n/status")
+                if str_chat_id in admin_ids:
+                    my_bot.sendMessage(chat_id=chat_id, text="/restart_nuc\n/start_ngrok\n/stop_ngrok\n/status")
+                else:
+                    my_bot.sendMessage(chat_id=chat_id, text="/status")
 
-
+            if no_rights:
+                my_bot.sendMessage(chat_id=chat_id, text="You don't have permission to run this command. " + emojize(':triumph:'))
 
     MessageLoop(my_bot, {'chat': on_chat_message}).run_as_thread()
     print('Listening ...')
